requirements.txt
aiogram>=2.0,<3.0
pandas
openpyxl
python-dotenv